+++ src/BinaryFile/MemoryMappedByteArrayFile.cs
@@ -5,12 +5,12 @@ namespace BinarySerializer
 {
     public class MemoryMappedByteArrayFile : MemoryMappedFile
     {
-        public MemoryMappedByteArrayFile(Context context, string name, uint baseAddress, uint length, Endian endianness = Endian.Little) : base(context, name, baseAddress, endianness)
+        public MemoryMappedByteArrayFile(Context context, string name, long baseAddress, long length, Endian endianness = Endian.Little) : base(context, name, baseAddress, endianness)
         {
             Bytes = new byte[length];
         }
 
-        public MemoryMappedByteArrayFile(Context context, string name, uint baseAddress, byte[] bytes, Endian endianness = Endian.Little) : base(context, name, baseAddress, endianness)
+        public MemoryMappedByteArrayFile(Context context, string name, long baseAddress, byte[] bytes, Endian endianness = Endian.Little) : base(context, name, baseAddress, endianness)
         {
             Bytes = bytes ?? throw new ArgumentNullException(nameof(bytes));
         }
@@ -37,7 +37,7 @@ namespace BinarySerializer
             return writer;
         }
 
-        public void WriteBytes(uint position, byte[] source) {
+        public void WriteBytes(long position, byte[] source) {
             Array.Copy(source, 0, Bytes, position, Math.Min(source.Length, Bytes.Length-position));
         }
